O45

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -67,7 +67,7 @@ namespace RestaurantAPI.Services
 
             return restaurant.Id;
         }
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Restaurant with Id: {id} DELETE action invoked");
             var restaurant = _dbContext
@@ -75,6 +75,14 @@ namespace RestaurantAPI.Services
                 .FirstOrDefault(r => r.Id == id);
             if(restaurant == null) 
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
